requirements.txt
azure-mgmt-resource
azure-mgmt-network
python-dotenv
azure-identity
azure-mgmt-compute